requirements.txt
azure-ai-inference
pydantic
python-dotenv
locust
